requirements.txt
torch
torchvision
albumentations
numpy
matplotlib
pandas
scikit-learn
Pillow
argparse